Bugfix in study to level order
When a unit with an AMTS has the order "study obse 5" it stops studying
as soon as it reaches OBSE-3. This is caused because void
Game::Do1StudyOrder() checks if the target level has been reached using
Unit::GetSkill(), but this function adds modifiers to the skill.

Unit::GetRealSkill() has to be used instead.